Move ErrorResponse to its own file.
This is part of the work for #1187.
