+++ google/cloud/storage/testbench/error_response.py
@@ -1,0 +1,36 @@
+#!/usr/bin/env python
+# Copyright 2018 Google LLC.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""A helper class to send error responses in the storage client test bench."""
+
+import flask
+
+
+class ErrorResponse(Exception):
+    """Simplify generation of error responses."""
+    status_code = 400
+
+    def __init__(self, message, status_code=None, payload=None):
+        Exception.__init__(self)
+        self.message = message
+        if status_code is not None:
+            self.status_code = status_code
+        self.payload = payload
+
+    def as_response(self):
+        kv = dict(self.payload or ())
+        kv['message'] = self.message
+        response = flask.jsonify(kv)
+        response.status_code = self.status_code
+        return response
