+++ google/cloud/storage/testbench/testbench.py
@@ -16,33 +16,15 @@
 
 import argparse
 import base64
-import json
-import time
 import flask
 import hashlib
 import httpbin
+import json
 import os
+import time
+from error_response import ErrorResponse
 from werkzeug import serving
 from werkzeug import wsgi
-
-
-class ErrorResponse(Exception):
-    """Simplify generation of error responses."""
-    status_code = 400
-
-    def __init__(self, message, status_code=None, payload=None):
-        Exception.__init__(self)
-        self.message = message
-        if status_code is not None:
-            self.status_code = status_code
-        self.payload = payload
-
-    def as_response(self):
-        kv = dict(self.payload or ())
-        kv['message'] = self.message
-        response = flask.jsonify(kv)
-        response.status_code = self.status_code
-        return response
 
 
 @httpbin.app.errorhandler(ErrorResponse)
